+++ test/rosserial_test/main/rosserial_test.cpp
@@ -21,7 +21,7 @@ void rosserial_setup(){  // Initialize ROS
   nh.subscribe(espSub);
 }
 
-void rosserial_publish(long long unsigned int* ticks_L, long long unsigned int* ticks_R){
+void rosserial_publish(long int* ticks_L, long int* ticks_R){
   ticks.motor_L = *ticks_L;
   ticks.motor_R = *ticks_R;
   espPub.publish(&ticks);  // publish the msg
